requirements.txt
# Coordinator Service Dependencies
# These are minimal dependencies for the central coordinator
# Each submodule has its own requirements.txt in its own venv

python-dotenv==1.0.1
requests==2.32.3
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.5
